requirements.txt
requests
beautifulsoup4 
pandas
boto3
python-dotenv